Ignore .idea packages in git repo. Renames packages in project from flichtinfo to flichtanalyzer
