+++ .gitignore
@@ -40,4 +40,5 @@ bin/
 .vscode/
 
 ### Mac OS ###
-.DS_Store
+.DS_Store.idea/encodings.xml
+.idea/misc.xml
