+++ src/main/java/org/flightanalyzer/domain/dto/FlightTicket.java
@@ -1,0 +1,136 @@
+package org.flightanalyzer.domain.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+
+/**
+ * This class represents a flight ticket with various details such as origin, destination,
+ * departure and arrival times, carrier, number of stops and price.
+ */
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class FlightTicket {
+    @JsonProperty("origin")
+    private String origin;
+
+    @JsonProperty("origin_name")
+    private String originName;
+
+    @JsonProperty("destination")
+    private String destination;
+
+    @JsonProperty("destination_name")
+    private String destinationName;
+
+    @JsonProperty("departure_date")
+    private String departureDate;
+
+    @JsonProperty("departure_time")
+    private String departureTime;
+
+    @JsonProperty("arrival_date")
+    private String arrivalDate;
+
+    @JsonProperty("arrival_time")
+    private String arrivalTime;
+
+    @JsonProperty("carrier")
+    private String carrier;
+
+    @JsonProperty("stops")
+    private int stops;
+
+    @JsonProperty("price")
+    private int price;
+
+    private List<FlightTicket> tickets;
+
+    public String getOrigin() {
+        return origin;
+    }
+
+    public void setOrigin(String origin) {
+        this.origin = origin;
+    }
+
+    public String getOriginName() {
+        return originName;
+    }
+
+    public void setOriginName(String originName) {
+        this.originName = originName;
+    }
+
+    public String getDestination() {
+        return destination;
+    }
+
+    public void setDestination(String destination) {
+        this.destination = destination;
+    }
+
+    public String getDestinationName() {
+        return destinationName;
+    }
+
+    public void setDestinationName(String destinationName) {
+        this.destinationName = destinationName;
+    }
+
+    public String getDepartureDate() {
+        return departureDate;
+    }
+
+    public void setDepartureDate(String departureDate) {
+        this.departureDate = departureDate;
+    }
+
+    public String getDepartureTime() {
+        return departureTime;
+    }
+
+    public void setDepartureTime(String departureTime) {
+        this.departureTime = departureTime;
+    }
+
+    public String getArrivalDate() {
+        return arrivalDate;
+    }
+
+    public void setArrivalDate(String arrivalDate) {
+        this.arrivalDate = arrivalDate;
+    }
+
+    public String getArrivalTime() {
+        return arrivalTime;
+    }
+
+    public void setArrivalTime(String arrivalTime) {
+        this.arrivalTime = arrivalTime;
+    }
+
+    public String getCarrier() {
+        return carrier;
+    }
+
+    public void setCarrier(String carrier) {
+        this.carrier = carrier;
+    }
+
+    public int getStops() {
+        return stops;
+    }
+
+    public void setStops(int stops) {
+        this.stops = stops;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public void setPrice(int price) {
+        this.price = price;
+    }
+}
