+++ src/main/java/org/flightanalyzer/domain/repository/TicketRepositoy.java
@@ -1,0 +1,9 @@
+package org.flightanalyzer.domain.repository;
+
+import org.flightanalyzer.domain.dto.FlightTicket;
+
+import java.util.List;
+
+public interface TicketRepositoy {
+    List<FlightTicket> getAllTickets();
+}
