+++ src/main/java/org/flightanalyzer/io/FlightDataReader.java
@@ -1,0 +1,9 @@
+package org.flightanalyzer.io;
+
+import org.flightanalyzer.domain.dto.FlightTicket;
+
+import java.util.List;
+
+public interface FlightDataReader {
+    List<FlightTicket> loadTickets();
+}
