--- src/main/java/org/flightinfo/domain/repository/TicketRepositoy.java
@@ -1,9 +1,0 @@
-package org.flightinfo.domain.repository;
-
-import org.flightinfo.domain.dto.FlightTicket;
-
-import java.util.List;
-
-public interface TicketRepositoy {
-    List<FlightTicket> getAllTickets();
-}
