--- src/main/java/org/flightinfo/io/FlightDataReader.java
@@ -1,9 +1,0 @@
-package org.flightinfo.io;
-
-import org.flightinfo.domain.dto.FlightTicket;
-
-import java.util.List;
-
-public interface FlightDataReader {
-    List<FlightTicket> loadTickets();
-}
