+++ src/test/java/org/flightanalyzer/service/FlightInfoServiceTest.java
@@ -1,0 +1,47 @@
+package org.flightanalyzer.service;
+
+import org.junit.jupiter.api.BeforeEach;
+
+import static org.mockito.Mockito.*;
+
+class FlightInfoServiceTest  {
+
+    private ConfigurationService cfgService;
+    private FlightInfoService flightInfoService;
+
+    @BeforeEach
+    public void setUp() {
+        cfgService = mock(ConfigurationService.class);
+        when(cfgService.getFlightFilePath()).thenReturn("/tickets.json");
+        flightInfoService = new FlightInfoService(cfgService);
+    }
+
+    /* @Test
+    public void testGetMinFlightTimes() {
+        Map<String, Duration> minFlightTimes = flightInfoService.getMinFlightTimes();
+        assertNotNull(minFlightTimes);
+        // Add more assertions based on expected values
+    }
+
+    @Test
+    public void testGetMinFlightTimesWithoutTimeZone() {
+        Map<String, Integer> minFlightTimesWithoutTimeZone = flightInfoService.getMinFlightTimesWithoutTimeZone();
+        assertNotNull(minFlightTimesWithoutTimeZone);
+        // Add more assertions based on expected values
+    }
+
+    @Test
+    public void testGetPriceDifference() {
+        double priceDifference = flightInfoService.getPriceDifference();
+        assertTrue(priceDifference >= 0);
+        // Add more assertions based on expected values
+    }
+
+   @Test
+    public void testProcessFlightData() {
+        flightInfoService.processFlightData(cfgService);
+        List<FlightTicket> tickets = flightInfoService.testGetTicketsUsingReflection();
+        assertNotNull(tickets);
+        // Add more assertions based on expected values
+    }*/
+}
